Move Dof stuff from /nodes to /dofs
