+++ src/mechanics/cell/Integrand.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "mechanics/cell/CellData.h"
 #include "mechanics/cell/CellIpData.h"
-#include "mechanics/nodes/DofVector.h"
-#include "mechanics/nodes/DofMatrix.h"
+#include "mechanics/dofs/DofVector.h"
+#include "mechanics/dofs/DofMatrix.h"
 #include "mechanics/cell/IPValue.h"
 
 namespace NuTo
