+++ test/mechanics/dofs/DofMatrix.cpp
@@ -1,5 +1,5 @@
 #include "BoostUnitTest.h"
-#include "mechanics/nodes/DofMatrix.h"
+#include "mechanics/dofs/DofMatrix.h"
 #include <sstream>
 
 NuTo::DofMatrix<int> Get(const NuTo::DofType& d0, const NuTo::DofType& d1)
