+++ test/mechanics/nodes/CMakeLists.txt
@@ -1,9 +1,3 @@
-add_unit_test(DofContainer)
-add_unit_test(DofVector)
-add_unit_test(DofMatrix)
-add_unit_test(DofMatrixSparse)
-add_unit_test(DofType)
-add_unit_test(NodeSimple)
 add_unit_test(NodeBase)
